fix: split test_client.py to satisfy pylint line limit
